+++ src/core.ts
@@ -17,7 +17,8 @@ export {
     PouchDB,
     validateCouchDBString,
     getBatch,
-    countAllUndeleted
+    countAllUndeleted,
+    getNewestSequence
 } from './pouch-db';
 
 export {