+++ test/unit.test.ts
@@ -1,7 +1,6 @@
 import './unit/init.test.js';
 import './unit/util.test.js';
 import './unit/pouch-db-integration.test.js';
-import './unit/attachments.test.js'; // TODO move down under plugin
 import './unit/rx-storage-pouchdb.test.js';
 import './unit/adapter-check.test.js';
 import './unit/instance-of-check.test.js';
@@ -26,6 +25,7 @@ import './unit/hooks.test.js';
 import './unit/orm.test.js';
 import './unit/population.test.js';
 import './unit/data-migration.test.js';
+import './unit/backup.test.js';
 import './unit/leader-election.test.js';
 import './unit/replication.test.js';
 import './unit/replication-graphql.test.js';
@@ -37,4 +37,5 @@ import './unit/local-documents.test.js';
 import './unit/in-memory.test.js';
 import './unit/server.test.js';
 import './unit/plugin.test.js';
+import './unit/attachments.test.js';
 import './unit/last.test.js';