+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### X.X.X (coming soon)
 
+Features:
+  - Added the [backup-plugin](https://rxdb.info/backup.html)
+
 ### 9.18.0 (26 April 2020)
 
 Bugfixes: